Remove unused 'stale' connection state and extract shared test FakeClient
The 'stale' value in ConnectionState was never assigned anywhere in the
state machine — staleness detection always transitioned directly to
'reconnecting'. This removes the dead type variant and the unreachable
check in queue-control-bar. Also extracts the duplicated FakeClient test
helper into a shared module.
