+++ packages/web/app/components/connection-manager/__tests__/websocket-connection-manager.test.ts
@@ -1,21 +1,6 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { connectionManager, STALE_GRACE_MS } from '../websocket-connection-manager';
-
-class FakeClient {
-  listeners = new Map<string, (...args: any[]) => void>();
-  terminate = vi.fn();
-  dispose = vi.fn();
-
-  on(event: string, listener: (...args: any[]) => void) {
-    this.listeners.set(event, listener);
-    return () => this.listeners.delete(event);
-  }
-
-  emit(event: string, ...args: any[]) {
-    const handler = this.listeners.get(event);
-    if (handler) handler(...args);
-  }
-}
+import { FakeClient } from './fake-client';
 
 describe('WebSocketConnectionManager', () => {
   let originalVisibilityState: PropertyDescriptor | undefined;
