+++ packages/web/app/components/connection-manager/__tests__/websocket-connection-provider.test.tsx
@@ -3,22 +3,7 @@ import { render, screen, act } from '@testing-library/react';
 import React from 'react';
 import { connectionManager } from '../websocket-connection-manager';
 import { WebSocketConnectionProvider, useWebSocketConnection } from '../websocket-connection-provider';
-
-class FakeClient {
-  listeners = new Map<string, (...args: any[]) => void>();
-  terminate = vi.fn();
-  dispose = vi.fn();
-
-  on(event: string, listener: (...args: any[]) => void) {
-    this.listeners.set(event, listener);
-    return () => this.listeners.delete(event);
-  }
-
-  emit(event: string, ...args: any[]) {
-    const handler = this.listeners.get(event);
-    if (handler) handler(...args);
-  }
-}
+import { FakeClient } from './fake-client';
 
 function TestConsumer() {
   const { state, name } = useWebSocketConnection();
