+++ packages/web/app/components/connection-manager/websocket-connection-manager.ts
@@ -1,6 +1,6 @@
 import { Client, Event as WsEvent, EventListener as WsEventListener } from 'graphql-ws';
 
-export type ConnectionState = 'idle' | 'connecting' | 'connected' | 'reconnecting' | 'stale' | 'error';
+export type ConnectionState = 'idle' | 'connecting' | 'connected' | 'reconnecting' | 'error';
 
 type ConnectionSnapshot = {
   name: string | null;
